# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.createVisitRequest:
    properties:
      category:
        type: string
      lat:
        type: number
      lng:
        type: number
      photo_reference:
        description: 画像参照（Redis TTL失効後の再解決に使用）
        type: string
      place_id:
        type: string
      place_name:
        type: string
      place_types:
        description: '任意: is_breakout自動判定に使用'
        items:
          type: string
        type: array
      user_lat:
        description: ユーザーの現在緯度（距離検証用）
        type: number
      user_lng:
        description: ユーザーの現在経度（距離検証用）
        type: number
      vicinity:
        type: string
      visited_at:
        type: string
    required:
    - category
    - lat
    - lng
    - place_id
    - place_name
    - visited_at
    type: object
  handlers.createVisitResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      daily_completed:
        description: 今回の訪問で本日の3件上限に達したか
        type: boolean
      genre_tag_id:
        type: integer
      id:
        type: integer
      is_breakout:
        type: boolean
      lat:
        type: number
      level_up:
        description: 今回の訪問でレベルアップしたか
        type: boolean
      lng:
        type: number
      memo:
        type: string
      new_badges:
        description: 今回獲得した新バッジ
        items:
          $ref: '#/definitions/models.Badge'
        type: array
      new_level:
        description: 現在のレベル
        type: integer
      photo_reference:
        type: string
      place_id:
        type: string
      place_name:
        type: string
      rating:
        type: number
      total_xp:
        description: ユーザー累計XP
        type: integer
      user_id:
        type: integer
      vicinity:
        type: string
      visited_at:
        type: string
      xp_breakdown:
        allOf:
        - $ref: '#/definitions/services.XPBreakdown'
        description: XP計算内訳
      xp_earned:
        type: integer
    type: object
...
  handlers.listVisitsResponse:
    properties:
      total:
        type: integer
      visits:
        items:
          $ref: '#/definitions/models.Visit'
        type: array
    type: object
...
  handlers.mapVisitItem:
    properties:
      category:
        type: string
      genre_tag_id:
        type: integer
      id:
        type: integer
      lat:
        type: number
      lng:
        type: number
      place_id:
        type: string
      place_name:
        type: string
      visited_at:
        type: string
    type: object
  handlers.mapVisitsResponse:
    properties:
      total:
        type: integer
      visits:
        items:
          $ref: '#/definitions/handlers.mapVisitItem'
        type: array
    type: object
...
  handlers.updateVisitRequest:
    properties:
      memo:
        type: string
      rating:
        type: number
    type: object
...
  models.Badge:
    properties:
      condition_json:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Visit:
    properties:
      category:
        type: string
      created_at:
        type: string
      genre_tag_id:
        type: integer
      id:
        type: integer
      is_breakout:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      memo:
        type: string
      photo_reference:
        type: string
      place_id:
        type: string
      place_name:
        type: string
      rating:
        type: number
      user_id:
        type: integer
      vicinity:
        type: string
      visited_at:
        type: string
      xp_earned:
        type: integer
    type: object
...
  services.XPBreakdown:
    properties:
      base_xp:
        description: ベースXP（通常50 or 脱却100）
        type: integer
      first_area_bonus:
        description: 初エリアボーナス
        type: integer
      streak_bonus:
        description: ストリークボーナス
        type: integer
    type: object
host: localhost:8000
info:
  contact: {}
  description: RoambleのバックエンドAPIドキュメント
  title: Roamble API
  version: "1.0"
paths:
...
  /api/visits:
    get:
      description: ユーザーの訪問履歴を一覧取得する（visited_at降順）
      parameters:
      - description: 取得件数（デフォルト20）
        in: query
        name: limit
        type: integer
      - description: オフセット（デフォルト0）
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.listVisitsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 訪問履歴取得
      tags:
      - Visits
    post:
      consumes:
      - application/json
      description: ユーザーの訪問記録を作成する（category、place_name を必須で受け付け）
      parameters:
      - description: 訪問記録データ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createVisitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.createVisitResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 訪問記録作成
      tags:
      - Visits
  /api/visits/{id}:
    get:
      description: 指定IDの訪問記録詳細を取得する（自分の記録のみ）
      parameters:
      - description: 訪問記録ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 訪問記録詳細取得
      tags:
      - Visits
    patch:
      consumes:
      - application/json
      description: ユーザーの訪問記録の感想メモ・評価を部分更新する
      parameters:
      - description: 訪問記録ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新データ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateVisitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Visit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 訪問記録更新
      tags:
      - Visits
  /api/visits/map:
    get:
      description: マップ表示に必要な位置情報・ジャンル情報を最適化して取得する
      parameters:
      - description: 取得件数（デフォルト・上限2000）
        in: query
        name: limit
        type: integer
      - description: オフセット（デフォルト0）
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.mapVisitsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: マップ表示用訪問データ取得
      tags:
      - Visits
...
securityDefinitions:
  BearerAuth:
    description: JWTアクセストークン（Bearer形式）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"